requirements.txt
python-telegram-bot>=20.8
flask[async]>=3.0.0
waitress>=3.0.0
supabase>=2.0.0
python-dotenv>=1.0.0